drivers: meida: pisp_be: Fix overflow in plane size calculations
The calculations for buffer plane sizes can overflow because of the
plane factor shift. Fix this by using u64 integers for the calculations.


(cherry picked from commit 0166e942456a64c5d9385afbcb69962c8fd45da8 pi5_61)

+++ drivers/media/platform/bcm2712/pisp_be.c
@@ -1092,7 +1092,7 @@ static void set_plane_params(struct v4l2_format *f,
 
 	for (i = 0; i < nplanes; i++) {
 		struct v4l2_plane_pix_format *p = &f->fmt.pix_mp.plane_fmt[i];
-		unsigned int bpl, plane_size;
+		u64 bpl, plane_size;
 
 		bpl = (f->fmt.pix_mp.width * fmt->bit_depth) >> 3;
 		bpl = ALIGN(max(p->bytesperline, bpl), fmt->align);
